conversion warning fix

+++ src/mtsmsme/comp/sri4smprint.cpp
@@ -37,6 +37,6 @@ int main(int argc, char* argv[])
     {
       SendRoutingInfoForSMReq* inv = new SendRoutingInfoForSMReq("79139859489", true, "79139869999");
       tsm->TInvokeReq(1, 45, *inv);
-      tsm->TBeginReq(sizeof(cd)/sizeof(uint8_t), cd, sizeof(cl)/sizeof(uint8_t), cl);
+      tsm->TBeginReq((uint8_t)(sizeof(cd)/sizeof(uint8_t)), cd, (uint8_t)(sizeof(cl)/sizeof(uint8_t)), cl);
     }
 }
